Create raing
based on csv file with 4 records (student1.csv) create a file rating.csv with the names of studets on budget and their avarage grade. The average_list array is updated, it only records the avarage grade of studets on budget, and doesnt write 0 according to non-budget students. Vector with the names of students on budget only is added as there is the need to write it to the rating.csv file.