Directory reader
dir_reader takes a path to folder (here on the desktop, but content od the folder is in files 1.csv and 2.csv) and returns a vector of strings with the names of files in the directory. There is also possibility for recursive reading of a directory. So the read_record function now reads from all files in the directory and create_rating works in an append mode