full functionality in test.cpp, results in rating2.cvs
But another version with student class, read_record2 creates students on the base of read lines from the file. The function  create_student parses the strings of data, validates them and if successful, creates a student. The newly created student entity is appended to student_list in read_record2 function