Student class implemented
There is a student.cpp and student.h library. Afteer reading each record, we create a student entite and add it to student_list, that for now is implemented as a vector (but there is probably a way to make it yet another class. Then we sort this student list by descending order, calculate the 40% of students on budget who will be in the list for scholarship and write those students to file rating1.csv