Change to vector of pointers
now functions that work with student list in a form of vector work with vector of student pointers. also dir_reader returns full path to each file. write_rating_to_file writes the average grade with precision of 3 numbres after the point (btw there is a little typo with the word 'average')